requirements.txt

langchain
langchain-core
langchain-openai
langgraph

pydantic>=2.0
typing_extensions

openai>=1.0.0
#deepseek

requests
httpx

python-dotenv

tqdm          # прогресс-бары (удобно при отладке)
tenacity      # для retry-логики (в LangChain используется под капотом, но лучше установить)